+++ packages/annotation/src/files/index.tsx
@@ -1,0 +1,3 @@
+export * from './Download'
+export * from './transformSvg'
+export * from './types'